requirements.txt
shiny @ git+https://github.com/posit-dev/py-shiny
python-dotenv
anthropic

# For emails
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
markdown
requests